backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
sqlalchemy[asyncio]>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
qdrant-client>=1.8
sentence-transformers>=2.6
pymupdf>=1.24
torch>=2.2